Reconcile/unify browser/native request-response testing
